Comparing an SVM with the SGD perceptron fit.
The SVM trained on the whole dataset. The SGD perceptron appears
to not find the separating hyperplanes even when they exist.

Edit: it's unclear what happened to the file. It may be that the version
of the file is mistakenly after the SGD perceptron code was removed
and replaced by using sklearn.linear_model.SGDClassifier.