Changed dataset to modified torchvision FakeData
Changed the random dataset using the torchvision FakeData dataset
as a guide. This allows for an rng to be used to draw random
imaged consistently according to an index, but doesn't
pollute the rng seed outside of the class for reasons I don't fully
understand. It's important to check to make sure labels are still
being generated randomly in a reasonable way. They seem to be.